move directory structures
